bundle scripts: corrections to copy openmpi script

+++ Bundlebot/nightly/copy_openmpi.sh
@@ -10,10 +10,15 @@ if [ ! -d $TODIR ]; then
   echo ***error: directory $TODIR does not exist
   exit
 fi
+if [ -d $TODIR/openmpi ]; then
+  rm -rf $TODIR/openmpi
+fi
 mkdir $TODIR/openmpi
 cd $FROMDIR
+if [ -e $TODIR/openmpi.tar ]; then
+  rm -f $TODIR/openmpi.tar
+fi
 tar cvf $TODIR/openmpi.tar .
 cd $TODIR/openmpi
-tar xvf ../$openmpi.tar
-rm -f $TODIR/openmpi.tar
-
+tar xvf $TODIR/openmpi.tar
+rm $TODIR/openmpi.tar
